final_project/requirements.txt
cv2
mtcnn
matplotlib
numpy
Pillow
scikit-learn
opencv-python-headless
streamlit
